showUserAccounts has been added.

+++ E-Banking/src/main/java/ebanking/model/entity/CurrentAccount.java
@@ -11,8 +11,8 @@ public class CurrentAccount extends Account {
 	private long currentAccountId;
 	private double creditLimit;
 
-	public CurrentAccount(long acountId, double netAvlbBalance, double currentBalance,
-			String iban, long userId, String currency, long currentAccountId, double creditLimit) 
+	public CurrentAccount(int acountId, double netAvlbBalance, double currentBalance,
+			String iban, int userId, String currency, int currentAccountId, double creditLimit) 
 					throws AccountException, IbanException, InvalidStringException, IdException {
 		super(acountId, netAvlbBalance, currentBalance, iban, userId, currency);
 
@@ -27,5 +27,5 @@ public class CurrentAccount extends Account {
 			throw new AccountException("Invalid credit limit");
 		}
 	}
-	
+
 }
